backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
django-storages
google-cloud-storage
python-decouple
Pillow
gunicorn
whitenoise
dj-database-url
django-filter
